requirements.txt
anyio==3.3.4
asyncclick==7.1.2.3
certifi==2021.10.8
charset-normalizer==2.0.7
colorthief==0.2.1
idna==3.3
importlib-metadata==4.8.1
Pillow==8.4.0
python-dotenv==0.19.1
python-kasa==0.4.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
spotipy==2.19.0
urllib3==1.26.7
zipp==3.6.0
